moved operator adders to their own file
