+++ src/index.ts
@@ -1,2 +1,1 @@
-export { proxify } from "./proxify-operator";
 export { createProxy } from "./create-proxy-observable";
